% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmark}
\alias{benchmark}
\title{benchmark}
\usage{
benchmark(index, seed)
}
\arguments{
\item{index}{Number of the function to be used.}

\item{seed}{Random seed for the evaluation.}
}
\description{
Function that calls the NMMSO algorithm with the 20 functions to be evaluated with.
}

