% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/niching_funcs.R
\name{CF1}
\alias{CF1}
\title{Composition Function 1}
\usage{
CF1(x)
}
\arguments{
\item{x}{Data matrix to be evaluated (2D).}
}
\value{
Result of the evaluation as a float number.
}
\description{
n = 6
}

