% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/niching_funcs.R
\name{CF2}
\alias{CF2}
\title{Composition Function 2}
\usage{
CF2(x)
}
\arguments{
\item{x}{Data matrix to be evaluated (2D).}
}
\value{
Result of the evaluation as a float number.
}
\description{
n = 8
}

