% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/niching_funcs.R
\name{CF3}
\alias{CF3}
\title{Composition Function 3}
\usage{
CF3(x)
}
\arguments{
\item{x}{Data matrix to be evaluated (2D, 3D, 5D, 10D).}
}
\value{
Result of the evaluation as a float number.
}
\description{
n = 6
}

