% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/niching_funcs.R
\name{CF4}
\alias{CF4}
\title{Composition Function 4}
\usage{
CF4(x)
}
\arguments{
\item{x}{Data matrix to be evaluated (3D, 5D, 10D, 20D).}
}
\value{
Result of the evaluation as a float number.
}
\description{
n = 8
}

