% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/count_goptima.R
\name{count_goptima}
\alias{count_goptima}
\title{count_goptima}
\usage{
count_goptima(pop, nfunc, accuracy)
}
\arguments{
\item{pop}{Data given to be evaluated.}

\item{nfunc}{Number of the niching function to be used.}

\item{accuracy}{Accuracy value for the optima selection criteria.}
}
\value{
A list containing the number of optimas found and the value of it. (count, finalseeds)
}
\description{
Evaluates the data given with a niching function and iterates it comparing the data with the rho parameter, determining if an optima has been found or not.
}

