% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/niching_funcs.R
\name{equal_maxima}
\alias{equal_maxima}
\title{F2: Equal Maxima}
\usage{
equal_maxima(x)
}
\arguments{
\item{x}{Data matrix to be evaluated (1D).}
}
\value{
Result of the evaluation as a float number.
}
\description{
Variables ranges: x in [0, 1]
No. of global peaks: 5
No. of local peaks: 0
}

