% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/niching_funcs.R
\name{F8F2}
\alias{F8F2}
\title{F8F2 Function}
\usage{
F8F2(x)
}
\arguments{
\item{x}{Data matrix to be evaluated.}
}
\value{
Result of the evaluation as a float number.
}
\description{
F8F2 Function
}

