% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/niching_funcs.R
\name{FEF8F2}
\alias{FEF8F2}
\title{FEF8F2 Function}
\usage{
FEF8F2(x)
}
\arguments{
\item{x}{Data matrix to be evaluated.}
}
\value{
Result of the evaluation as a float number.
}
\description{
FEF8F2 Function
}

