% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/feval.R
\name{feval}
\alias{feval}
\title{feval}
\usage{
feval(f, ...)
}
\arguments{
\item{f}{Function given to be run.}

\item{...}{Passing arguments on to other functions.}
}
\description{
Helper functions which imitates the behavior of the Matlab feval.
}

