% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/niching_funcs.R
\name{FGriewank}
\alias{FGriewank}
\title{Griewank's Function}
\usage{
FGriewank(x)
}
\arguments{
\item{x}{Data matrix to be evaluated.}
}
\value{
Result of the evaluation as a float number.
}
\description{
Griewank's Function
}

