% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/niching_funcs.R
\name{five_uneven_peak_trap}
\alias{five_uneven_peak_trap}
\title{F1: Five-Uneven-Peak Trap}
\usage{
five_uneven_peak_trap(x)
}
\arguments{
\item{x}{Data matrix to be evaluated (1D).}
}
\value{
Result of the evaluation as a float number.
}
\description{
Variables ranges: x in [0, 30]
No. of global peaks: 2
No. of local peaks: 3
}

