% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/niching_funcs.R
\name{FWeierstrass}
\alias{FWeierstrass}
\title{Weierstrass Function}
\usage{
FWeierstrass(x)
}
\arguments{
\item{x}{Data matrix to be evaluated.}
}
\value{
Result of the evaluation as a float number.
}
\description{
Weierstrass Function
}

