% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_fgoptima.R
\name{get_fgoptima}
\alias{get_fgoptima}
\title{get_fgoptima}
\usage{
get_fgoptima(nfunc)
}
\arguments{
\item{nfunc}{Number of the niching function to be used.}
}
\value{
The global optimal value for the function needed.
}
\description{
get_fgoptima
}

