% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_no_goptima.R
\name{get_no_goptima}
\alias{get_no_goptima}
\title{get_no_goptima}
\usage{
get_no_goptima(nfunc)
}
\arguments{
\item{nfunc}{Number of the niching function to be used.}
}
\value{
The number of global optimal values to be found depending the function.
}
\description{
get_no_goptima
}

