% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_rho.R
\name{get_rho}
\alias{get_rho}
\title{get_rho}
\usage{
get_rho(nfunc)
}
\arguments{
\item{nfunc}{Number of the niching function to be used.}
}
\value{
The rho value for the function needed.
}
\description{
get_rho
}

