% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_ub.R
\name{get_ub}
\alias{get_ub}
\title{get_ub}
\usage{
get_ub(fno)
}
\arguments{
\item{fno}{Number of the niching function to be used.}
}
\value{
The upper bound for the function needed.
}
\description{
get_ub
}

