% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/niching_funcs.R
\name{himmelblau}
\alias{himmelblau}
\title{F4: Himmelblau}
\usage{
himmelblau(x)
}
\arguments{
\item{x}{Data matrix to be evaluated (2D).}
}
\value{
Result of the evaluation as a float number.
}
\description{
Variables ranges: x, y in [-6, 6]
No. of global peaks: 4
No. of local peaks: 0
}

