% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/niching_funcs.R
\name{hybrid_composition_func}
\alias{hybrid_composition_func}
\title{Hybrid Composition General Framework}
\usage{
hybrid_composition_func(x, func_num, func, o, sigma, lambda, bias, M,
  CF3_CF4_flag)
}
\arguments{
\item{x}{Data matrix to be evaluated.}

\item{func_num}{Number of the function to be used.}

\item{func}{Function to use during the evaluation.}

\item{o}{Omega parameter (Matlab's data file containing a data matrix)}

\item{sigma}{Vector containing different sigma values.}

\item{lambda}{Vector containing different lambda values.}

\item{bias}{Matrix 1xfunc_num full with 0s.}

\item{M}{Identity matrix.}
}
\value{
Result of the evaluation as a float number.
}
\description{
Hybrid Composition General Framework
}

