% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/niching_funcs.R
\name{local_gram_schmidt}
\alias{local_gram_schmidt}
\title{Classical Gram Schmidt Function}
\usage{
local_gram_schmidt(A)
}
\arguments{
\item{x}{Data matrix to be evaluated (3D, 5D, 10D, 20D).}
}
\value{
Resulting matrix with the q information.
}
\description{
Classical Gram Schmidt Function
}

