% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/niching_funcs.R
\name{modified_rastrigin_all}
\alias{modified_rastrigin_all}
\title{F8: Modified Rastrigin - All Global Optima}
\usage{
modified_rastrigin_all(x)
}
\arguments{
\item{x}{Data matrix to be evaluated (2D).}
}
\value{
Result of the evaluation as a float number.
}
\description{
Variables ranges: x_i in [0, 1]^n, i = 1, 2, ..., n
No. of global peaks: \prod_{i=1}^n k_i
No. of local peaks: 0
}

