% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/niching_funcs.R
\name{niching_func}
\alias{niching_func}
\title{niching_func}
\usage{
niching_func(x, func_num)
}
\arguments{
\item{x}{is a 1xD input vector for evaluation.}

\item{func_num}{denotes the number of the objective function which is going to be used.}
}
\value{
The objective function value of the x input vector.
}
\description{
Benchmark Functions for CEC'2013 Special Session and Competition on Niching Methods for Multimodal Function Optimization.
This benchmark set includes the following 12 multimodal test functions:
F1 : Five-Uneven-Peak Trap (1D)
F2 : Equal Maxima (1D)
F3 : Uneven Decreasing Maxima (1D)
F4 : Himmelblau (2D)
F5 : Six-Hump Camel Back (2D)
F6 : Shubert (2D, 3D)
F7 : Vincent (2D, 3D)
F8 : Modified Rastrigin - All Global Optima (2D)
F9 : Composition Function 1 (2D)
F10 : Composition Function 2 (2D)
F11 : Composition Function 3 (2D, 3D, 5D, 10D)
F12 : Composition Function 4 (3D, 5D, 10D, 20D)

For more information please refer to the Technical Report of the
Special Session/Competition at: http://goanna.cs.rmit.edu.au/~xiaodong/cec13-niching/
}

