% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/niching_funcs.R
\name{rot_matrix_condition}
\alias{rot_matrix_condition}
\title{Rotation Matrix}
\usage{
rot_matrix_condition(D, c)
}
\arguments{
\item{D}{Number of dimensions for the matrix.}

\item{c}{Constant matrix 1xfunc_num full of 1s.}
}
\description{
Generates a D-dimensional rotation matrix with predifined Condition Number (c)
}

