% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/niching_funcs.R
\name{shubert}
\alias{shubert}
\title{F6: Shubert}
\usage{
shubert(x)
}
\arguments{
\item{x}{Data matrix to be evaluated (2D, 3D).}
}
\value{
Result of the evaluation as a float number.
}
\description{
Variables ranges: x_i in [-10, 10]^n, i = 1, 2, ..., n
No. of global peaks: n * 3^n
No. of local peaks: many
}

