% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/niching_funcs.R
\name{six_hump_camel_back}
\alias{six_hump_camel_back}
\title{F5: Six-Hump Camel Back}
\usage{
six_hump_camel_back(x)
}
\arguments{
\item{x}{Data matrix to be evaluated (2D).}
}
\value{
Result of the evaluation as a float number.
}
\description{
Variables ranges: x in [-1.9, 1.9]; y in [-1.1, 1.1]
No. of global peaks: 2
No. of local peaks: 2
}

