% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/niching_funcs.R
\name{uneven_decreasing_maxima}
\alias{uneven_decreasing_maxima}
\title{F3: Uneven Decreasing Maxima}
\usage{
uneven_decreasing_maxima(x)
}
\arguments{
\item{x}{Data matrix to be evaluated (1D).}
}
\value{
Result of the evaluation as a float number.
}
\description{
Variables ranges: x in [0, 1]
No. of global peaks: 1
No. of local peaks: 4
}

